+++ app/javascript/mastodon/locales/defaultMessages.json
@@ -829,10 +829,6 @@
   {
     "descriptors": [
       {
-        "defaultMessage": "Column settings",
-        "id": "notifications.settings"
-      },
-      {
         "defaultMessage": "Desktop notifications",
         "id": "notifications.column_settings.alert"
       },
@@ -883,16 +879,21 @@
   {
     "descriptors": [
       {
-        "defaultMessage": "Notifications",
-        "id": "column.notifications"
-      },
-      {
         "defaultMessage": "Are you sure you want to permanently clear all your notifications?",
         "id": "notifications.clear_confirmation"
       },
       {
         "defaultMessage": "Clear notifications",
         "id": "notifications.clear"
+      }
+    ],
+    "path": "app/javascript/mastodon/features/notifications/containers/column_settings_container.json"
+  },
+  {
+    "descriptors": [
+      {
+        "defaultMessage": "Notifications",
+        "id": "column.notifications"
       },
       {
         "defaultMessage": "You don't have any notifications yet. Interact with others to start the conversation.",
