+++ app/javascript/mastodon/locales/id.json
@@ -107,7 +107,6 @@
   "notifications.column_settings.reblog": "Boost:",
   "notifications.column_settings.show": "Tampilkan dalam kolom",
   "notifications.column_settings.sound": "Mainkan suara",
-  "notifications.settings": "Pengaturan kolom",
   "onboarding.done": "Selesei",
   "onboarding.next": "Selanjutnya",
   "onboarding.page_five.public_timelines": "Linimasa lokal menampilkan semua postingan publik dari semua orang di {domain}. Linimasa gabungan menampilkan postingan publik dari semua orang yang diikuti oleh {domain}. Ini semua adalah Linimasa Publik, cara terbaik untuk bertemu orang lain.",
